requirements.txt
streamlit
python-dotenv
langchain
langchain_core
langchain_openai
langchain_community